pyproject.toml
[project]
name = "repl-nix-therapyyybot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.55.3",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot>=21.7",
]
